fix deleting review thats not from shorediving

+++ app/__init__.py
@@ -592,7 +592,8 @@ def delete_review():
   for image in review.images:
     Image.query.filter_by(id=image.id).delete()
 
-  ShoreDivingReview.query.filter_by(id=review.shorediving_data.id).delete()
+  if review.shorediving_data:
+    ShoreDivingReview.query.filter_by(id=review.shorediving_data.id).delete()
 
   spot = Spot.query.filter_by(id=beach_id).first()
   if spot.num_reviews == 1:
